requirements/base.txt
# Core dependencies for SubSkin project
# Data collection and processing
metapub>=0.5.0          # PubMed API library
requests>=2.20.0       # HTTP library
beautifulsoup4>=4.10.0 # HTML parsing
lxml>=4.8.0            # XML parsing for PubMed

# AI processing
openai>=1.0.0          # OpenAI API client
anthropic>=0.20.0      # Anthropic API client (optional alternative)
python-dotenv>=1.0.0   # Environment variable management

# Data processing and storage
pandas>=1.5.0          # Data manipulation
pydantic>=2.0.0        # Data validation and settings management
sqlalchemy>=2.0.0      # ORM for database
psycopg2-binary>=2.9.0 # PostgreSQL adapter (for future use)
redis>=4.0.0           # Redis client for caching

# Web framework (for API backend)
fastapi>=0.100.0       # Modern, fast web framework
uvicorn[standard]>=0.20.0 # ASGI server
jinja2>=3.0.0          # Templating engine
httpx>=0.20.0          # Async HTTP client

# Utilities
python-dateutil>=2.8.0 # Date parsing
tqdm>=4.60.0           # Progress bars for long-running tasks
pyyaml>=6.0            # YAML configuration parsing
tenacity>=8.0.0        # Retry library for API calls

# Documentation generation
mkdocs>=1.5.0          # Documentation site generator
mkdocs-material>=9.0   # Material theme for documentation

# Alibaba Cloud VISION AI SDK (for VASI assessment)
alibabacloud-python-sdk-core>=1.0.0